UI visual upgrades, aded fractions to stock and etf investments, updated readme

+++ backend/src/main/java/eye/on/the/money/service/stock/RSUGrantService.java
@@ -180,7 +180,7 @@ public class RSUGrantService implements ICSVService {
                 .exchange(grant.getExchange())
                 .currency(grant.getCurrency())
                 .date(tranche.vestDate().isAfter(today) ? today : tranche.vestDate())
-                .quantity(tranche.quantity())
+                .quantity((double) tranche.quantity())
                 .build();
     }
